Some internal renames/refactorings
The previous file layout was mostly a historic artifact, due to having
two callback implementations. Since we have dropped the "slow" callback,
it makes sense ro restructure things a bit.
