+++ src/oop_ext/foundation/callback/__init__.py
@@ -1,5 +1,5 @@
 from ._callbacks import Callbacks
-from ._fast_callback import Callback
+from ._callback import Callback
 from ._priority_callback import PriorityCallback
 from ._shortcuts import After, Before, Remove, WrapForCallback
